Labfiles/04b-autogen/python/requirements.txt
python-dotenv
azure-identity
python-dotenv
autogen-agentchat
autogen-ext[azure]
aiohttp